requirements.txt
aiogram==3.6.0
aiohttp==3.9.3
tortoise-orm==0.20.0
telethon==1.34.0
pydantic==2.6.4
python-dotenv==1.0.1
upstash-redis==1.2.0
requests==2.31.0
aiofiles==23.2.1
asyncio-throttle==1.0.2